+++ src/Util/Configuration/Logging/CodeCoverage/Crap4j.php
@@ -7,26 +7,39 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace PHPUnit\Util\Configuration;
+namespace PHPUnit\Util\Configuration\Logging\CodeCoverage;
+
+use PHPUnit\Util\Configuration\File;
 
 /**
  * @internal This class is not covered by the backward compatibility promise for PHPUnit
  * @psalm-immutable
  */
-final class TestDoxXml
+final class Crap4j
 {
     /**
      * @var File
      */
     private $target;
 
-    public function __construct(File $target)
+    /**
+     * @var int
+     */
+    private $threshold;
+
+    public function __construct(File $target, int $threshold)
     {
-        $this->target = $target;
+        $this->target    = $target;
+        $this->threshold = $threshold;
     }
 
     public function target(): File
     {
         return $this->target;
+    }
+
+    public function threshold(): int
+    {
+        return $this->threshold;
     }
 }
