+++ src/Util/Configuration/Logging/CodeCoverage/Html.php
@@ -7,13 +7,15 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace PHPUnit\Util\Configuration;
+namespace PHPUnit\Util\Configuration\Logging\CodeCoverage;
+
+use PHPUnit\Util\Configuration\Directory;
 
 /**
  * @internal This class is not covered by the backward compatibility promise for PHPUnit
  * @psalm-immutable
  */
-final class CodeCoverageHtml
+final class Html
 {
     /**
      * @var Directory
