+++ src/Util/Configuration/Logging/CodeCoverage/Php.php
@@ -7,13 +7,15 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace PHPUnit\Util\Configuration;
+namespace PHPUnit\Util\Configuration\Logging\CodeCoverage;
+
+use PHPUnit\Util\Configuration\File;
 
 /**
  * @internal This class is not covered by the backward compatibility promise for PHPUnit
  * @psalm-immutable
  */
-final class CodeCoveragePhp
+final class Php
 {
     /**
      * @var File
