+++ src/Util/Configuration/Logging/CodeCoverage/Xml.php
@@ -7,25 +7,27 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace PHPUnit\Util\Configuration;
+namespace PHPUnit\Util\Configuration\Logging\CodeCoverage;
+
+use PHPUnit\Util\Configuration\Directory;
 
 /**
  * @internal This class is not covered by the backward compatibility promise for PHPUnit
  * @psalm-immutable
  */
-final class CodeCoverageClover
+final class Xml
 {
     /**
-     * @var File
+     * @var Directory
      */
     private $target;
 
-    public function __construct(File $target)
+    public function __construct(Directory $target)
     {
         $this->target = $target;
     }
 
-    public function target(): File
+    public function target(): Directory
     {
         return $this->target;
     }
