+++ src/Util/Configuration/Logging/TestDox/Text.php
@@ -7,25 +7,27 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace PHPUnit\Util\Configuration;
+namespace PHPUnit\Util\Configuration\Logging\TestDox;
+
+use PHPUnit\Util\Configuration\File;
 
 /**
  * @internal This class is not covered by the backward compatibility promise for PHPUnit
  * @psalm-immutable
  */
-final class CodeCoverageXml
+final class Text
 {
     /**
-     * @var Directory
+     * @var File
      */
     private $target;
 
-    public function __construct(Directory $target)
+    public function __construct(File $target)
     {
         $this->target = $target;
     }
 
-    public function target(): Directory
+    public function target(): File
     {
         return $this->target;
     }
